Implemented basic routing and folder structure for redux toolkit
