+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import logo from './logo.svg';
-import { Counter } from './features/counter/Counter';
+import { Counter } from './components/Recipe';
 import './App.css';
 
 function App() {
