+++ src/components/Recipe.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { decrement, increment } from '../reducers/recipeSlice';
+
+export function Counter() {
+  const count = useSelector((state) => state.recipe.value);
+  const dispatch = useDispatch();
+
+  return (
+    <div>
+      <div>
+        <button
+          aria-label="Increment value"
+          onClick={() => dispatch(increment())}
+        >
+          Increment
+        </button>
+        <span>{count}</span>
+        <button
+          aria-label="Decrement value"
+          onClick={() => dispatch(decrement())}
+        >
+          Decrement
+        </button>
+      </div>
+    </div>
+  );
+}
