+++ src/reducers/recipeSlice.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-export const counterSlice = createSlice({
+export const recipeSlice = createSlice({
   name: 'counter',
   initialState: {
     value: 0,
@@ -11,10 +11,10 @@ export const counterSlice = createSlice({
       // doesn't actually mutate the state because it uses the Immer library,
       // which detects changes to a "draft state" and produces a brand new
       // immutable state based off those changes
-      state.value += 1;
+      state.value = 'hola';
     },
     decrement: (state) => {
-      state.value -= 1;
+      state.value = 'adios';
     },
     incrementByAmount: (state, action) => {
       state.value += action.payload;
@@ -23,6 +23,6 @@ export const counterSlice = createSlice({
 });
 
 // Action creators are generated for each case reducer function
-export const { increment, decrement, incrementByAmount } = counterSlice.actions;
+export const { increment, decrement, incrementByAmount } = recipeSlice.actions;
 
-export default counterSlice.reducer;
+export default recipeSlice.reducer;
